Made Song.GetInstrument's return value non-nullable. Changed the fg colour argument to TrackViewWindowBase.DrawChannelHeader to a byte. Added new utility type ByReferenceList<T> for structs that returns ref T from the indexer. Added additional constructors and operators to EnvelopeNode.cs. Translated song_get_playing_instruments and song_update_playing_instrument into corresponding methods in Song.cs. Refactored SongNote.NoteString, exposing a static SongNote.GetNoteString for external use. Added additional operators to Size.cs. Updated the DrawText family of methodn in VGAMem.cs to return the number of characters drawn. Added helper method Widget.ContainsPoint. Renamed InstrumentListGeneralPage.cs to InstrumentListGeneralSubpage.cs. Renamed InstrumentListVolumePage.cs to InstrumentListVolumeSubpage.cs. Renamed InstrumentListPanningPage.cs to InstrumentListPanningSubpage.cs. Renamed InstrumentListPitchPage.cs to InstrumentListPitchSubpage.cs. Introduced intermediate class InstrumentListEnvelopeSubpageBase.cs that is a common base for the subpages displaying envelopes. Added stubs for ITI.cs and XI.cs to FileTypes/Converters. Translated page_instruments.c into InstrumentListPage.cs, InstrumentListEnvelopeSubpageBase.cs, InstrumentListGeneralSubpage.cs, InstrumentListVolumeSubpage.cs, InstrumentListPanningSubpage.cs, InstrumentListPitchSubpage.cs, EnvelopeResizeDialog.cs, EnvelopeADSRDialog.cs and ExportInstrumentDialog.cs. Fixed up old code in MiniPopSlideDialog.cs that was written before the VGAMem interface was finalized. Added using statements.

+++ Pages/InfoWindows/TrackView13Window.cs
@@ -15,7 +15,7 @@ public class TrackView13Window : TrackViewWindowBase
 	protected override int ChannelWidth => 13;
 	protected override bool Separator => true;
 
-	protected override void DrawChannelHeader(int chan, int column, int @base, int fg)
+	protected override void DrawChannelHeader(int chan, int column, int @base, byte fg)
 	{
 		TrackView.DrawChannelHeader(chan, new Point(column, @base), fg);
 	}
