+++ Pages/InfoWindows/TrackView7Window.cs
@@ -17,7 +17,7 @@ public class TrackView7Window : TrackViewWindowBase
 	protected override int ChannelWidth => 7;
 	protected override bool Separator => false;
 
-	protected override void DrawChannelHeader(int chan, int column, int @base, int fg)
+	protected override void DrawChannelHeader(int chan, int column, int @base, byte fg)
 	{
 		VGAMem.DrawCharacter(0, new Point(column + 0, @base), (1, 1));
 		VGAMem.DrawCharacter(0, new Point(column + 1, @base), (1, 1));
